perf: separate some components.
