+++ package.json
@@ -31,6 +31,7 @@
     "@tamagui/theme-base": "^1.14.3",
     "@types/react": "~18.0.27",
     "babel-plugin-module-resolver": "^5.0.0",
+    "crossbell.js": "^0.19.8",
     "dotenv": "^16.0.3",
     "expo": "~48.0.10",
     "expo-dev-client": "~2.1.6",
