+++ src/pages/ArticleList/constants.ts
@@ -1,0 +1,5 @@
+export enum SortType {
+    LATEST = 'latest',
+    POPULAR = 'popular',
+    FOLLOW = 'follow'
+}
